Move code and clean up code
